# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Cantina da Sandra",
    "description": "API para o sistema de pedidos da cantina da Sandra"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/clientes/login": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para realizar login de um cliente.",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "description": "Login do cliente.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Login": {
                  "type": "string",
                  "example": "string"
                },
                "Senha": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "Login",
                "Senha"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "cliente": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string",
                      "example": "number"
                    },
                    "Nome": {
                      "type": "string",
                      "example": "string"
                    },
                    "Login": {
                      "type": "string",
                      "example": "string"
                    },
                    "Senha": {
                      "type": "string",
                      "example": "string"
                    },
                    "Ativado": {
                      "type": "string",
                      "example": "boolean"
                    },
                    "CPF": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                },
                "token": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Login ou senha inválidos."
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para obter um cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do cliente."
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado.",
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "string",
                  "example": "number"
                },
                "Nome": {
                  "type": "string",
                  "example": "string"
                },
                "Login": {
                  "type": "string",
                  "example": "string"
                },
                "Senha": {
                  "type": "string",
                  "example": "string"
                },
                "Ativado": {
                  "type": "string",
                  "example": "boolean"
                },
                "CPF": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Cliente não encontrado."
          }
        }
      }
    },
...
    "/clientes/cadastrar": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para cadastrar um cliente.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do cliente.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Nome": {
                  "type": "string",
                  "example": "string"
                },
                "Login": {
                  "type": "string",
                  "example": "string"
                },
                "Senha": {
                  "type": "string",
                  "example": "string"
                },
                "CPF": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "Nome",
                "Login",
                "Senha",
                "CPF"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente cadastrado com sucesso."
          },
          "500": {
            "description": "Erro ao cadastrar o cliente."
          }
        }
      }
    },
    "/clientes/desativar/{id}": {
      "delete": {
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para desativar um cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "description": "Token de autenticação.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente desativado com sucesso."
          },
          "404": {
            "description": "Cliente não encontrado."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
...
}